Fixed: throw error for unknows CLI flags
