+++ __tests__/cli.js
@@ -1,31 +1,39 @@
 import test from "ava"
 import { join } from "path"
 
-import { denodeify as asyncify } from "promise"
 import { exec } from "child_process"
-
-const asyncExec = asyncify(exec)
 
 const target = join(__dirname, "..", "test-boilerplate")
 const execOpts = { cwd: target }
 
 const statinamic = "./node_modules/.bin/statinamic"
 
-test("should throw if a CLI flag is NOT recognized", async (t) => {
-  return (
-    asyncExec(`${ statinamic } start --lol`, execOpts)
-    .then(
-      () => t.fail(),
-      (err) => {
-        t.ok(err.message.indexOf("Unknown option `--lol'") > -1)
+// cannot be done for now
+// see the mess in configurator which use process.argv directly
+// and since build.js is spawned, we got too many args and cannot use
+// yargs.strict()
+test.cb("should throw if a CLI flag is NOT recognized", (t) => {
+  const child = exec(
+    `${ statinamic } start --lol --open=false`, execOpts,
+    (err) => {
+      if (err) {
+        clearTimeout(timeout)
+        t.ok(err.message.indexOf("Unknown argument") > -1)
+        t.end()
       }
-    )
+    }
   )
+
+  const timeout = setTimeout(() => {
+    child.kill()
+    t.fail()
+    t.end()
+  }, 500)
 })
 
 test.cb("should NOT throw if a CLI flag is recognized", (t) => {
   const child = exec(
-    `${ statinamic } start --devPort=4000`, execOpts,
+    `${ statinamic } start --devPort=4000 --open=false`, execOpts,
 
     // should die quickly...
     (err) => {
@@ -37,11 +45,9 @@ test.cb("should NOT throw if a CLI flag is recognized", (t) => {
     }
   )
 
-  // ...or be ok quickly
-  // (so we assume it's ok and kill the process, we don't need the actual build)
   const timeout = setTimeout(() => {
+    child.kill()
     t.pass()
     t.end()
-    child.kill()
   }, 500)
 })
